BY3 move one frame into separate directory
