+++ forex-mtl/src/main/scala/forex/thirdPartyApi/oneFrame/errors.scala
@@ -1,4 +1,4 @@
-package forex.thirdPartyApi
+package forex.thirdPartyApi.oneFrame
 
 object errors {
